requirements.txt
flask>=3.0
psycopg2-binary>=2.9
flask-login>=0.6
bcrypt>=4.0
flask-limiter>=3.5
gunicorn>=22.0
python-dotenv>=1.0
